requirements.txt
blinker==1.9.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.8
colorzero==2.0
distro==1.8.0
Flask==3.1.0
gpiod==2.2.3
gpiodevice==0.0.5
gpiozero==2.0.1
gunicorn==23.0.0
idna==3.3
inky==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
lgpio==0.2.2.0
MarkupSafe==3.0.2
numpy==2.2.2
packaging==24.2
pdoc==15.0.1
pigpio==1.78
pillow==11.1.0
prometheus_client==0.21.1
psutil==6.1.1
pycryptodomex==3.11.0
Pygments==2.19.1
pytest==8.1.1
pytest-cov==4.1.0
pytest-mock==3.12.0
python-apt==2.6.0
requests==2.28.1
rpi-lgpio==0.6
six==1.16.0
smbus2==0.4.2
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
urllib3==1.26.12
Werkzeug==3.1.3
